New port: net-p2p/nicotine-plus
Nicotine-Plus is a fork of Hyriand's original Nicotine Soulseek client.

Nicotine+ is an attempt to keep Nicotine working with the latest libraries,
kill bugs, keep current with the Soulseek protocol and add some new features
that users want and/or need. The main developer now is daelstorm.

WWW:    http://nicotine-plus.sourceforge.net/