The Gambit programming system is a full implementation of the Scheme language which conforms to the R4RS and IEEE Scheme standards. It consists of two main programs: gsi, the Gambit Scheme interpreter, and gsc, the Gambit Scheme compiler.
Gambit-C is a version of the Gambit programming system in which the
compiler generates portable C code, making the whole Gambit-C system
and the programs compiled with it easily portable to many computer
architectures for which a C compiler is available. With appropriate
declarations in the source code the executable programs generated by
the compiler run roughly as fast as equivalent C programs.

WWW: http://www.iro.umontreal.ca/~gambit/

PR:		ports/107405
Submitted by:	Rob Zinkov