Add port multimedia/gaupol:
Gaupol is an editor for text-based subtitle files. It supports multiple
subtitle file formats and provides convenient means of correcting texts
and timing subtitles to match video. The user interface is designed with
attention to batch processing of multiple documents and convenience of
translating.

WWW: http://home.gna.org/gaupol/
Author: Osmo Salomaa <otsaloma@cc.hut.fi>