Update to new shiny version 0.5.0.
We have a new fast, robust and flexible pathfinder, a brandnew terrain
generator, full UNICODE support and better multiplayer.  NewGRF
compatibility (modding) has been greatly enhanced with support for changing
stations, adding sounds, more vehicle sets, etc.

The smaller feature- and buglist goes on and on, but we'll let the user
explore this.  All users of 0.4.8 are encouraged to update to this new
release.  Enjoy!