Add port graphics/tesseract:
The Tesseract OCR engine was one of the top 3 engines in the 1995 UNLV
Accuracy test. Between 1995 and 2006 it had little work done on it, but
it is probably one of the most accurate open source OCR engines
available. The source code will read a binary, grey or color image and
output text.

WWW: http://code.google.com/p/tesseract-ocr/

Based on:	pkgsrc package