Add port misc/mtail:
mtail is a small tail workalike that performs output coloring using ansi
escape sequences (although the sequences are overridable, so you could cause
it to output something else, e.g. html font tags, if you really wanted to).
mtail is written in python, is fairly small, and should be relatively
platform-independent.

It has a config file that can contain an arbitrary number of entries, each
of which has a series of regular expressions to indicate which files to color
according to which entry. for each entry, the config file specifies a coloring
scheme using regular expressions and, optionally, filters to apply to each
line before coloring (for example, to strip out extra info, etc.). the config
file also may override the predefined colors and the escape sequences (or
whatever) actually used to perform the coloring.

WWW: http://matt.immute.net/src/mtail/
Author: Matt Hellige <matt@immute.net>

Based on:	NetBSD pkgsrc package