Begin the de-orbit burn for the beloved, but now more than a bit long in the tooth BIND 8. As of today (27 August 2007) ISC has announced that BIND 8 is officially End of Life (EOL) and therefore it's time to say good-bye.
Please see http://www.isc.org/sw/bind/bind8-eol.php for details on the
reasoning behind the EOL status, the latest security issues, and a
migration guide to help you move toward BIND 9.4.x.

bind8 (BIND 8.3.7) is marked FORBIDDEN due to the predictable query ID
bug (see above) which will not be fixed for this version.

bind84 is marked DEPRECATED, and will be upgraded to 8.4.7-P1 when it
is available.