- Take maintainership
HTML::FormFu is a HTML form framework which aims to be as easy as possible to use for basic
web forms, but with the power and flexibility to do anything else you might want to do
(as long as it involves forms).

You can configure almost any part of formfu's behaviour and output. By default formfu renders
"XHTML 1.0 Strict" compliant markup, with as little extra markup as possible, but with
sufficient CSS class names to allow for a wide-range of output styles to be generated by
changing only the CSS.

PR:		ports/118994
Submitted by:	az@