OpenBSD's Relay Daemon (previously known as hoststated).
relayd is a daemon to relay and dynamically redirect incoming
connections to a target host. Its main purposes are to run as a
load-balancer, application layer gateway, or transparent proxy. The
daemon is able to monitor groups of hosts for availability, which is
determined by checking for a specific service common to a host group.

WWW: http://spootnik.org/relayd/

# This port will work on $OSVERSION >= 700049.
# If you want to use on RELENG_6, apply a patch in
# http://www.openbsd.org/cgi-bin/cvsweb/src/sys/net/pf_table.c.diff?r1=1.67&r2=1.68