This is a statistic plugin for SqueezeCenter. SqueezeCenter normally stores statistics about the last time a track was played, when a specific track was added to the library and how many times a track has been played. The problem is that all of these statistics are cleared every time you perform a full rescan of SqueezeCenter.
The TrackStat plugin solves this problem by making sure that the
statistics survive a rescan. Besides this TrackStat also extends the
statistics a bit, one example is that it doesn't count a track as
played just because you listen to the first 2 seconds of it, you will
have to play a certain amount of the track until it is played.
TrackStat also makes it possible to put a rating on all your tracks by
holding a number between 1-5 down on the now playing screen.

The other main functionality the TrackStat plugins provides is various
ways to browse your music based on the statistic information. The
standard SqueezeCenter only makes it possible to show most played
tracks. The TrackStat plugin makes it possible to show the statistics
in a lot more ways. The purpose is simply to give you another way to
select which music you like to play, for example like:

    * Find tracks you haven't played for a long time
    * Find tracks you have recently added to the library
    * Find top rated tracks
    * Find least played tracks
    * And a lot more...

WWW: http://wiki.slimdevices.com/index.php/TrackStat_plugin