Add port math/octave-forge-triangular.
This is triangular.
Simple example of a user type implementing a simple matrix type
for triangular matrices.

The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

PR:             127030
Submitted by:   Stephen Montgomery-Smith <stephen@math.missouri.edu>