Nathive is a libre software image editor, similar to Adobe Photoshop, Corel Photo-Paint or The GIMP, but focused on usability, logic and providing a smooth learning curve for everyone. The project run in the Gnome desktop environment and anyone is welcome to collaborate on it with code, translations or ideas.
This project is in the alpha phase, so it is an incomplete work,
unfit for the end user yet. The intention is to achieve a professional
graphic editor progressively without giving up initial usability.
Nathive is written from scratch in C using GTK+, and is designed
to be simple, lightweight, and easy to install and use.

WWW: http://www.nathive.org/